requirements.txt
altgraph==0.17.4
bcrypt==4.0.1
bitarray==2.9.2
bitstring==4.1.4
bitstruct==8.19.0
bleak==0.22.2
blinker==1.8.2
branca==0.7.2
Cartopy==0.23.0
cbor2==5.6.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dnspython==2.6.1
EasySettings==4.0.1
ecdsa==0.18.0
esp-idf-panic-decoder==1.1.0
esptool==4.8.1
et-xmlfile==1.1.0
Flask==3.0.3
folium==0.16.0
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
iniconfig==2.0.0
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mplcursors==0.5.3
mysql-connector-python==8.3.0
Naked==0.1.32
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pefile==2023.2.7
pexpect==4.9.0
pillow==10.2.0
pluggy==1.5.0
ptyprocess==0.7.0
pyarrow==15.0.0
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pyelftools==0.31
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pymongo==4.7.3
pynput==1.7.6
pyparsing==3.1.1
pypng==0.20220715.0
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
pyshp==2.3.1
pysmb==1.2.9.1
pytest==8.2.2
pytest-embedded==1.10.3
pytest-embedded-idf==1.10.3
pytest-embedded-serial==1.10.3
pytest-embedded-serial-esp==1.10.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyVISA==1.14.1
PyVISA-py==0.7.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzbar==0.1.9
qasync==0.27.1
qrcode==7.4.2
QtPy==2.4.1
qtwidgets==1.1
reedsolo==1.7.0
requests==2.31.0
scikit-learn==1.5.0
scipy==1.14.0
setuptools==69.1.0
shapely==2.0.4
shellescape==3.8.1
six==1.16.0
smmap==5.0.1
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
websockets==12.0
Werkzeug==3.0.3
winrt-runtime==2.1.0
winrt-Windows.Devices.Bluetooth==2.1.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.1.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.1.0
winrt-Windows.Devices.Enumeration==2.1.0
winrt-Windows.Foundation==2.1.0
winrt-Windows.Foundation.Collections==2.1.0
winrt-Windows.Storage.Streams==2.1.0
xyzservices==2024.4.0
zebra==0.1.0
